Properly open in terminal multibuffer excerpts
